requirements.txt
Flask==3.0.0
mysql-connector-python>=9.0.0
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.16.2
protobuf==4.25.3
mediapipe==0.10.21
